Introduce command line argument recreateDb.
Recreate Db only in development mode

+++ Program.cs
@@ -4,51 +4,54 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using RazorMvc.Data;
 using System;
+using System.Linq;
 
 namespace RazorMvc
 {
     public class Program
     {
-        public static void Main(string[] args) 
+        public static void Main(string[] args)
         {
         var host = CreateHostBuilder(args).Build();
-
-        CreateDbIfNotExists(host);
+        bool recreateDb = args.Contains("--recreateDb");
+        InitializeDb(host, recreateDb);
 
         host.Run();
         }
 
-        private static void CreateDbIfNotExists(IHost host)
-        {
-        using (var scope = host.Services.CreateScope())
-        {
-            var services = scope.ServiceProvider;
-            try
-            {
-                var context = services.GetRequiredService<InternDbContext>();
-                var webHostEnvironment = services.GetRequiredService< IWebHostEnvironment>();
-                if (webHostEnvironment.IsDevelopment())
-                    {
-                        context.Database.EnsureDeleted();
-                        context.Database.EnsureCreated();
-                    }
-
-                SeedData.Initialize(context);
-            }
-            catch (Exception ex)
-            {
-                var logger = services.GetRequiredService<ILogger<Program>>();
-                logger.LogError(ex, "An error occurred creating the DB.");
-            }
-        }
-    }
-
-
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     webBuilder.UseStartup<Startup>();
                 });
+
+        private static void InitializeDb(IHost host, bool recreateDb)
+        {
+            using (var scope = host.Services.CreateScope())
+            {
+                var services = scope.ServiceProvider;
+                var logger = services.GetRequiredService<ILogger<Program>>();
+
+                try
+                {
+                    var context = services.GetRequiredService<InternDbContext>();
+                    var webHostEnvironment = services.GetRequiredService<IWebHostEnvironment>();
+                    if (webHostEnvironment.IsDevelopment() && recreateDb)
+                        {
+                            logger.LogDebug("User requested to recreate Database");
+                            context.Database.EnsureDeleted();
+                            context.Database.EnsureCreated();
+                            logger.LogWarning("The Database was recreated.");
+                        }
+
+                    SeedData.Initialize(context);
+                }
+                catch (Exception ex)
+                {
+                    logger.LogError(ex, "An error occurred creating the DB.");
+                }
+            }
+        }
     }
 }
